@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """This module contains a storage class and codec to support the N5 format.
 """
 import os
@@ -97,7 +96,7 @@ class N5Store(NestedDirectoryStore):
 
             key = invert_chunk_coords(key)
 
-        return super(N5Store, self).__getitem__(key)
+        return super().__getitem__(key)
 
     def __setitem__(self, key, value):
 
@@ -144,7 +143,7 @@ class N5Store(NestedDirectoryStore):
 
             key = invert_chunk_coords(key)
 
-        super(N5Store, self).__setitem__(key, value)
+        super().__setitem__(key, value)
 
     def __delitem__(self, key):
 
@@ -157,7 +156,7 @@ class N5Store(NestedDirectoryStore):
         elif is_chunk_key(key):
             key = invert_chunk_coords(key)
 
-        super(N5Store, self).__delitem__(key)
+        super().__delitem__(key)
 
     def __contains__(self, key):
 
@@ -184,7 +183,7 @@ class N5Store(NestedDirectoryStore):
 
             key = invert_chunk_coords(key)
 
-        return super(N5Store, self).__contains__(key)
+        return super().__contains__(key)
 
     def __eq__(self, other):
         return (
@@ -200,7 +199,7 @@ class N5Store(NestedDirectoryStore):
         # We can't use NestedDirectoryStore's listdir, as it requires
         # array_meta_key to be present in array directories, which this store
         # doesn't provide.
-        children = super(NestedDirectoryStore, self).listdir(path=path)
+        children = super().listdir(path=path)
 
         if self._is_array(path):
 
@@ -244,7 +243,7 @@ class N5Store(NestedDirectoryStore):
 
     def _load_n5_attrs(self, path):
         try:
-            s = super(N5Store, self).__getitem__(path)
+            s = super().__getitem__(path)
             return json_loads(s)
         except KeyError:
             return {}
@@ -305,13 +304,15 @@ def invert_chunk_coords(key):
 def group_metadata_to_n5(group_metadata):
     '''Convert group metadata from zarr to N5 format.'''
     del group_metadata['zarr_format']
+    # TODO: This should only exist at the top-level
     group_metadata['n5'] = '2.0.0'
     return group_metadata
 
 
 def group_metadata_to_zarr(group_metadata):
     '''Convert group metadata from N5 to zarr format.'''
-    del group_metadata['n5']
+    # This only exists at the top level
+    group_metadata.pop('n5', None)
     group_metadata['zarr_format'] = ZARR_FORMAT
     return group_metadata
 
@@ -460,7 +461,7 @@ def compressor_config_to_n5(compressor_config):
 
     else:  # pragma: no cover
 
-        raise RuntimeError("Unknown compressor with id %s" % codec_id)
+        n5_config.update({k: v for k, v in compressor_config.items() if k != 'type'})
 
     return n5_config
 
@@ -512,7 +513,7 @@ def compressor_config_to_zarr(compressor_config):
 
     else:  # pragma: no cover
 
-        raise RuntimeError("Unknown compressor with id %s" % codec_id)
+        zarr_config.update({k: v for k, v in compressor_config.items() if k != 'type'})
 
     return zarr_config
 
