@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """This module contains storage classes for use with Zarr arrays and groups.
 
 Note that any object implementing the :class:`MutableMapping` interface from the
@@ -29,19 +28,23 @@ import warnings
 import zipfile
 from collections import OrderedDict
 from collections.abc import MutableMapping
-from os import scandir, replace
+from os import scandir
 from pickle import PicklingError
 from threading import Lock, RLock
 from typing import Optional, Union, List, Tuple, Dict
 import uuid
 import time
 
-from numcodecs.compat import ensure_bytes, ensure_contiguous_ndarray
+from numcodecs.compat import (
+    ensure_bytes,
+    ensure_text,
+    ensure_contiguous_ndarray
+)
 from numcodecs.registry import codec_registry
 
 from zarr.errors import (
     MetadataError,
-    err_bad_compressor,
+    BadCompressorError,
     ContainsArrayError,
     ContainsGroupError,
     FSPathExistNotDir,
@@ -56,7 +59,7 @@ from zarr.meta import (
 )
 from zarr.util import (buffer_size, json_loads, nolock, normalize_chunks,
                        normalize_dtype, normalize_fill_value, normalize_order,
-                       normalize_shape, normalize_storage_path)
+                       normalize_shape, normalize_storage_path, retry_call)
 
 __doctest_requires__ = {
     ('RedisStore', 'RedisStore.*'): ['redis'],
@@ -437,8 +440,8 @@ def _init_array_metadata(
     if compressor:
         try:
             compressor_config = compressor.get_config()
-        except AttributeError:
-            err_bad_compressor(compressor)
+        except AttributeError as e:
+            raise BadCompressorError(compressor) from e
 
     # obtain filters config
     if filters:
@@ -765,7 +768,7 @@ class DictStore(MemoryStore):
                       "will be removed in the future. Please use MemoryStore.",
                       DeprecationWarning,
                       stacklevel=2)
-        super(DictStore, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
 
 class DirectoryStore(MutableMapping):
@@ -912,8 +915,10 @@ class DirectoryStore(MutableMapping):
         try:
             self._tofile(value, temp_path)
 
-            # move temporary file into place
-            replace(temp_path, file_path)
+            # move temporary file into place;
+            # make several attempts at writing the temporary file to get past
+            # potential antivirus file locking issues
+            retry_call(os.replace, (temp_path, file_path), exceptions=(PermissionError,))
 
         finally:
             # clean up if temp file still exists for whatever reason
@@ -1072,27 +1077,37 @@ class FSStore(MutableMapping):
     storage_options : passed to the fsspec implementation
     """
 
+    _META_KEYS = (attrs_key, group_meta_key, array_meta_key)
+
     def __init__(self, url, normalize_keys=True, key_separator='.',
                  mode='w',
                  exceptions=(KeyError, PermissionError, IOError),
                  **storage_options):
         import fsspec
-        self.path = url
         self.normalize_keys = normalize_keys
         self.key_separator = key_separator
         self.map = fsspec.get_mapper(url, **storage_options)
         self.fs = self.map.fs  # for direct operations
+        self.path = self.fs._strip_protocol(url)
         self.mode = mode
         self.exceptions = exceptions
-        if self.fs.exists(url) and not self.fs.isdir(url):
+        if self.fs.exists(self.path) and not self.fs.isdir(self.path):
             raise FSPathExistNotDir(url)
 
     def _normalize_key(self, key):
         key = normalize_storage_path(key).lstrip('/')
         if key:
             *bits, end = key.split('/')
-            key = '/'.join(bits + [end.replace('.', self.key_separator)])
+
+            if end not in FSStore._META_KEYS:
+                end = end.replace('.', self.key_separator)
+                key = '/'.join(bits + [end])
+
         return key.lower() if self.normalize_keys else key
+
+    def getitems(self, keys, **kwargs):
+        keys = [self._normalize_key(key) for key in keys]
+        return self.map.getitems(keys, on_error="omit")
 
     def __getitem__(self, key):
         key = self._normalize_key(key)
@@ -1101,16 +1116,22 @@ class FSStore(MutableMapping):
         except self.exceptions as e:
             raise KeyError(key) from e
 
+    def setitems(self, values):
+        if self.mode == 'r':
+            raise ReadOnlyError()
+        values = {self._normalize_key(key): val for key, val in values.items()}
+        self.map.setitems(values)
+
     def __setitem__(self, key, value):
         if self.mode == 'r':
             raise ReadOnlyError()
         key = self._normalize_key(key)
         path = self.dir_path(key)
-        value = ensure_contiguous_ndarray(value)
         try:
             if self.fs.isdir(path):
                 self.fs.rm(path, recursive=True)
             self.map[key] = value
+            self.fs.invalidate_cache(self.fs._parent(path))
         except self.exceptions as e:
             raise KeyError(key) from e
 
@@ -1194,7 +1215,7 @@ class TempStore(DirectoryStore):
     def __init__(self, suffix='', prefix='zarr', dir=None, normalize_keys=False):
         path = tempfile.mkdtemp(suffix=suffix, prefix=prefix, dir=dir)
         atexit.register(atexit_rmtree, path)
-        super(TempStore, self).__init__(path, normalize_keys=normalize_keys)
+        super().__init__(path, normalize_keys=normalize_keys)
 
 
 _prog_ckey = re.compile(r'^(\d+)(\.\d+)+$')
@@ -1283,23 +1304,23 @@ class NestedDirectoryStore(DirectoryStore):
     """
 
     def __init__(self, path, normalize_keys=False):
-        super(NestedDirectoryStore, self).__init__(path, normalize_keys=normalize_keys)
+        super().__init__(path, normalize_keys=normalize_keys)
 
     def __getitem__(self, key):
         key = _nested_map_ckey(key)
-        return super(NestedDirectoryStore, self).__getitem__(key)
+        return super().__getitem__(key)
 
     def __setitem__(self, key, value):
         key = _nested_map_ckey(key)
-        super(NestedDirectoryStore, self).__setitem__(key, value)
+        super().__setitem__(key, value)
 
     def __delitem__(self, key):
         key = _nested_map_ckey(key)
-        super(NestedDirectoryStore, self).__delitem__(key)
+        super().__delitem__(key)
 
     def __contains__(self, key):
         key = _nested_map_ckey(key)
-        return super(NestedDirectoryStore, self).__contains__(key)
+        return super().__contains__(key)
 
     def __eq__(self, other):
         return (
@@ -1308,7 +1329,7 @@ class NestedDirectoryStore(DirectoryStore):
         )
 
     def listdir(self, path=None):
-        children = super(NestedDirectoryStore, self).listdir(path=path)
+        children = super().listdir(path=path)
         if array_meta_key in children:
             # special handling of directories containing an array to map nested chunk
             # keys back to standard chunk keys
@@ -1619,18 +1640,6 @@ def migrate_1to2(store):
     del store['attrs']
 
 
-def _dbm_encode_key(key):
-    if hasattr(key, 'encode'):
-        key = key.encode('ascii')
-    return key
-
-
-def _dbm_decode_key(key):
-    if hasattr(key, 'decode'):
-        key = key.decode('ascii')
-    return key
-
-
 # noinspection PyShadowingBuiltins
 class DBMStore(MutableMapping):
     """Storage class using a DBM-style database.
@@ -1760,7 +1769,10 @@ class DBMStore(MutableMapping):
             with self.write_mutex:
                 if hasattr(self.db, 'sync'):
                     self.db.sync()
-                else:
+                else:  # pragma: no cover
+                    # we don't cover this branch anymore as ndbm (oracle) is not packaged
+                    # by conda-forge on non-mac OS:
+                    # https://github.com/conda-forge/staged-recipes/issues/4476
                     # fall-back, close and re-open, needed for ndbm
                     flag = self.flag
                     if flag[0] == 'n':
@@ -1776,17 +1788,20 @@ class DBMStore(MutableMapping):
         self.close()
 
     def __getitem__(self, key):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         return self.db[key]
 
     def __setitem__(self, key, value):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         value = ensure_bytes(value)
         with self.write_mutex:
             self.db[key] = value
 
     def __delitem__(self, key):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         with self.write_mutex:
             del self.db[key]
 
@@ -1800,7 +1815,7 @@ class DBMStore(MutableMapping):
         )
 
     def keys(self):
-        return (_dbm_decode_key(k) for k in iter(self.db.keys()))
+        return (ensure_text(k, "ascii") for k in iter(self.db.keys()))
 
     def __iter__(self):
         return self.keys()
@@ -1809,18 +1824,9 @@ class DBMStore(MutableMapping):
         return sum(1 for _ in self.keys())
 
     def __contains__(self, key):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         return key in self.db
-
-
-def _lmdb_decode_key_buffer(key):
-    # assume buffers=True
-    return key.tobytes().decode('ascii')
-
-
-def _lmdb_decode_key_bytes(key):
-    # assume buffers=False
-    return key.decode('ascii')
 
 
 class LMDBStore(MutableMapping):
@@ -1912,10 +1918,6 @@ class LMDBStore(MutableMapping):
         self.db = lmdb.open(path, **kwargs)
 
         # store properties
-        if buffers:
-            self.decode_key = _lmdb_decode_key_buffer
-        else:
-            self.decode_key = _lmdb_decode_key_bytes
         self.buffers = buffers
         self.path = path
         self.kwargs = kwargs
@@ -1947,7 +1949,8 @@ class LMDBStore(MutableMapping):
         self.close()
 
     def __getitem__(self, key):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         # use the buffers option, should avoid a memory copy
         with self.db.begin(buffers=self.buffers) as txn:
             value = txn.get(key)
@@ -1956,18 +1959,21 @@ class LMDBStore(MutableMapping):
         return value
 
     def __setitem__(self, key, value):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         with self.db.begin(write=True, buffers=self.buffers) as txn:
             txn.put(key, value)
 
     def __delitem__(self, key):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         with self.db.begin(write=True) as txn:
             if not txn.delete(key):
                 raise KeyError(key)
 
     def __contains__(self, key):
-        key = _dbm_encode_key(key)
+        if isinstance(key, str):
+            key = key.encode("ascii")
         with self.db.begin(buffers=self.buffers) as txn:
             with txn.cursor() as cursor:
                 return cursor.set_key(key)
@@ -1976,13 +1982,13 @@ class LMDBStore(MutableMapping):
         with self.db.begin(buffers=self.buffers) as txn:
             with txn.cursor() as cursor:
                 for k, v in cursor.iternext(keys=True, values=True):
-                    yield self.decode_key(k), v
+                    yield ensure_text(k, "ascii"), v
 
     def keys(self):
         with self.db.begin(buffers=self.buffers) as txn:
             with txn.cursor() as cursor:
                 for k in cursor.iternext(keys=True, values=False):
-                    yield self.decode_key(k)
+                    yield ensure_text(k, "ascii")
 
     def values(self):
         with self.db.begin(buffers=self.buffers) as txn:
@@ -2288,6 +2294,7 @@ class ABSStore(MutableMapping):
 
     def __contains__(self, key):
         blob_name = self._append_path_to_prefix(key)
+        assert len(blob_name) >= 1
         if self.client.exists(self.container, blob_name):
             return True
         else:
@@ -2312,6 +2319,7 @@ class ABSStore(MutableMapping):
         if dir_path:
             dir_path += '/'
         for blob in self.client.list_blobs(self.container, prefix=dir_path):
+            assert len(blob.name) >= 1
             self.client.delete_blob(self.container, blob.name)
 
     def getsize(self, path=None):
@@ -2320,9 +2328,11 @@ class ABSStore(MutableMapping):
         fs_path = self.prefix
         if store_path:
             fs_path = self._append_path_to_prefix(store_path)
-        if self.client.exists(self.container, fs_path):
-            return self.client.get_blob_properties(self.container,
-                                                   fs_path).properties.content_length
+
+        if fs_path != "" and self.client.exists(self.container, fs_path):
+            return self.client.get_blob_properties(
+                self.container, fs_path
+            ).properties.content_length
         else:
             size = 0
             if fs_path == '':
