@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from collections.abc import MutableMapping
 from itertools import islice
 
@@ -1098,9 +1097,11 @@ class Group(MutableMapping):
         self._write_op(self._move_nosync, source, dest)
 
 
-def _normalize_store_arg(store, clobber=False, storage_options=None):
-    return normalize_store_arg(store, clobber=clobber, default=MemoryStore,
-                               storage_options=storage_options)
+def _normalize_store_arg(store, *, clobber=False, storage_options=None, mode=None):
+    if store is None:
+        return MemoryStore()
+    return normalize_store_arg(store, clobber=clobber,
+                               storage_options=storage_options, mode=mode)
 
 
 def group(store=None, overwrite=False, chunk_store=None,
@@ -1209,8 +1210,10 @@ def open_group(store=None, mode='a', cache_attrs=True, synchronizer=None, path=N
     """
 
     # handle polymorphic store arg
-    clobber = mode != 'r'
-    store = _normalize_store_arg(store, clobber=clobber, storage_options=storage_options)
+    clobber = mode != "r"
+    store = _normalize_store_arg(
+        store, clobber=clobber, storage_options=storage_options, mode=mode
+    )
     if chunk_store is not None:
         chunk_store = _normalize_store_arg(chunk_store, clobber=clobber,
                                            storage_options=storage_options)
