@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import base64
 import json
 from collections.abc import Mapping
@@ -9,7 +8,7 @@ from zarr.errors import MetadataError
 from zarr.util import json_dumps, json_loads
 import zarr.util
 
-from typing import Union, Any, List, Mapping as MappingType
+from typing import cast, Union, Any, List, Mapping as MappingType
 
 ZARR_FORMAT = 2
 ZARR_FORMAT_v3 = "3"
@@ -255,8 +254,9 @@ def encode_fill_value(v: Any, dtype: np.dtype) -> Any:
     elif dtype.kind == 'b':
         return bool(v)
     elif dtype.kind in 'c':
-        v = (encode_fill_value(v.real, dtype.type().real.dtype),
-             encode_fill_value(v.imag, dtype.type().imag.dtype))
+        c = cast(np.complex128, np.dtype(complex).type())
+        v = (encode_fill_value(v.real, c.real.dtype),
+             encode_fill_value(v.imag, c.imag.dtype))
         return v
     elif dtype.kind in 'SV':
         v = str(base64.standard_b64encode(v), 'ascii')
