@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
 import atexit
 import os
 import pickle
 import shutil
-import sys
 import tempfile
 import textwrap
 import unittest
@@ -1077,8 +1075,7 @@ class TestGroupWithZipStore(TestGroup):
 
         # Check that exiting the context manager closes the store,
         # and therefore the underlying ZipFile.
-        error = ValueError if sys.version_info >= (3, 6) else RuntimeError
-        with pytest.raises(error):
+        with pytest.raises(ValueError):
             store.zf.extractall()
 
 
@@ -1364,14 +1361,14 @@ def test_tree():
     g5.create_dataset('baz', shape=100, chunks=10)
 
     # test root group
-    expect_bytes = textwrap.dedent(u"""\
+    expect_bytes = textwrap.dedent("""\
     /
      +-- bar
      |   +-- baz
      |   +-- quux
      |       +-- baz (100,) float64
      +-- foo""").encode()
-    expect_text = textwrap.dedent(u"""\
+    expect_text = textwrap.dedent("""\
     /
      ├── bar
      │   ├── baz
@@ -1381,19 +1378,19 @@ def test_tree():
     _check_tree(g1, expect_bytes, expect_text)
 
     # test different group
-    expect_bytes = textwrap.dedent(u"""\
+    expect_bytes = textwrap.dedent("""\
     foo""").encode()
-    expect_text = textwrap.dedent(u"""\
+    expect_text = textwrap.dedent("""\
     foo""")
     _check_tree(g2, expect_bytes, expect_text)
 
     # test different group
-    expect_bytes = textwrap.dedent(u"""\
+    expect_bytes = textwrap.dedent("""\
     bar
      +-- baz
      +-- quux
          +-- baz (100,) float64""").encode()
-    expect_text = textwrap.dedent(u"""\
+    expect_text = textwrap.dedent("""\
     bar
      ├── baz
      └── quux
